chg: test code for auction
